Add files via upload
The purpose of this script is to make a connection with a MySQL database and that offers 3 options to the users after they typed the right password. Only those users can do that who has given permissions to the MySQL's connection. You can feel free to use it, you are free to change and redistribute it.

I know this script is not the perfect one, but while I was writing the lines, every moment was so exciting as I wrote my first real shell script which is usable for a real task.